+++ src/lexocraft/llm/vector_database.hpp
@@ -20,8 +20,7 @@ namespace lc {
 
         using Vector_t = Eigen::Vector<float, WORD_VECTOR_DIMENSIONS>;
 
-        static MAPBOX_ETERNAL_CONSTEXPR const auto SOUNDEX_CODES =
-            mapbox::eternal::map<char, float>({
+        static MAPBOX_ETERNAL_CONSTEXPR const auto SOUNDEX_CODES = mapbox::eternal::map<char, int>({
   // clang-format off
                 {'b', 1}, {'f', 1}, {'p', 1}, {'v', 1},
                 {'c', 2}, {'g', 2}, {'j', 2}, {'k', 2}, {'q', 2}, {'s', 2}, {'x', 2},
@@ -44,7 +43,7 @@ namespace lc {
         std::string word;
         Eigen::Vector<float, WORD_VECTOR_DIMENSIONS> vector;
 
-        [[nodiscard]] std::vector<float> soundex() const;
+        [[nodiscard]] std::vector<int> soundex() const;
         [[nodiscard]] float soundex_distance(const WordVector& other) const;
         [[nodiscard]] float levenshtein_distance(const WordVector& other) const;
         [[nodiscard]] float similarity(const WordVector& other, float soundex_weight = 0.5F,
